Can now change the size of the grid by using a slider
The createGrid() function is now createGridCells(), we use it
to create the grid cells of an already existing grid with default dims
when the user opens the website. Plus we have added a new function
createNewGrid() to create and return a new whole grid with the size specified by the user.
This function is used by the updateGrid() function which replaces
the existing grid by the new one created by createNewGrid().
This function itself is called every time the user modifies the value of the slider.